# Rebatch after filters when using pipeline.ordered