--- a/logstash-core/src/main/java/org/logstash/config/ir/PipelineIR.java
+++ b/logstash-core/src/main/java/org/logstash/config/ir/PipelineIR.java
@@ -28,6 +28,7 @@ import org.logstash.config.ir.graph.Graph;
 import org.logstash.config.ir.graph.PluginVertex;
 import org.logstash.config.ir.graph.QueueVertex;
 import org.logstash.config.ir.graph.Vertex;
+import org.logstash.config.ir.graph.SeparatorVertex;
 
 public final class PipelineIR implements Hashable {
 
@@ -42,7 +43,9 @@ public final class PipelineIR implements Hashable {
     }
 
     private final Graph graph;
+
     private final QueueVertex queue;
+
     // Temporary until we have LIR execution
     // Then we will no longer need this property here
     private final String originalSource;
@@ -62,6 +65,9 @@ public final class PipelineIR implements Hashable {
 
         // Now we connect the queue to the root of the filter section
         tempGraph = tempGraph.chain(filterSection);
+
+        // Connect the filter section to the filter end vertex to separate from the output section
+        tempGraph = tempGraph.chain(new SeparatorVertex("filter_to_output"));
 
         // Finally, connect the filter out node to all the outputs
         this.graph = tempGraph.chain(outputSection);