Fixes #11550
Relates to #11175

#### WIP TODOs
- [x] major cleanup
- [x] refactor & simplify the `filtersExecution` + `lastFilter` logic
- [x] re-enable and fix relevant java tests
- [x] benchmark

#### Description
When using `pipeline.ordered` the batch events are sent 1 by 1 down the filters and output. This PR rebatches after the filters so that full batches can be sent to the outputs. It does that by separating the filters and output execution with the introduction of a separator vertex in the pipeline graph between the filters and outputs and compiling both sections separately and in the worker_loop calling both compute separately which provides the possibility to rebatch after the filters.