@colinsurprenant 2020-03-20
Unrelated build failure I restarted the failed build and it is now green https://logstash-ci.elastic.co/job/elastic+logstash+pull-request+multijob-integration-1/1161/