@colinsurprenant 2020-03-23
I performed the same benchmarks as in https://github.com/elastic/logstash/pull/11524#issuecomment-578220606 (but on different hardware, but it's not really important since relative numbers are interesting here).  Also, I noticed variations of ~5-10% between runs with the same parameters which I attribute to CPU throttling. 

## Config1
#### Java Execution

pipeline.ordered |  TPS master | TPS PR
------------- | ------------- | -------------:
false | 158k/s | 156k/s
true | 80k/s | **155k/s**

#### Ruby Execution

pipeline.ordered | TPS master | TPS PR
------------- | ------------- | -------------:
false/true  | 152k/s | 157k/s

## Config2
#### Java Execution

pipeline.ordered |  TPS master | TPS PR
------------- | ------------- | -------------:
false | 70k/s | 75k/s
true | 38k/s | **56k/s**

#### Ruby Execution

pipeline.ordered | TPS master | TPS PR
------------- | ------------- | -------------:
false/true  | 60k/s | 63k/s