@colinsurprenant 2020-03-25
Thanks a lot for the reviews @yaauie @andsel for the reviews!  I pushed some suggested modifications, will be working on others today. I'd also like your thoughts about the `SeparatorVertex` comments (keep as-is with an explicit ID or refactor into one more abstract class). LMKWYT.