@colinsurprenant 2020-03-26
@andsel @yaauie I think we are ready for a another round of review - as per my comments 
- I suggest we leave the `SeparatorVertex` implementation as-is here and followup with a broader `Vertex` refactor which will include the `QueueVertex`.
- Also defer a `AckedReadBatch` refactor. There is some cleanups and simplifications to be done there but is out of scope here.