@colinsurprenant 2020-03-30
Merged in master (8.0), 7.x #11726 and 7.7 #11730